refactor: extract 5 event modules from MediaLive.Show
Extract EpisodeEvents, DownloadEvents, MediaItemEvents, CategoryEvents,
and CollectionEvents into focused modules under show/. Each module
exposes public functions matching event names, dispatched from thin
one-liner handle_event clauses in show.ex.

Also moves monitoring_preset_label, available_categories_for, and
media_type_path to shared helpers for reuse across event modules and
components.
